Twilio-XILabs Voice Synthesizer - Python/requirements.txt
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
boto3==1.34.8
botocore==1.34.8
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
distro==1.9.0
elevenlabs==0.2.27
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.106.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
ipython==8.19.0
jedi==0.19.1
jmespath==1.0.1
matplotlib-inline==0.1.6
multidict==6.0.4
mypy-extensions==1.0.0
openai==1.6.1
packaging==23.2
parso==0.8.3
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.5.3
pydantic_core==2.14.6
pydub==0.25.1
Pygments==2.17.2
PyJWT==2.8.0
python-dateutil==2.8.2
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
starlette==0.27.0
tomli==2.0.1
tqdm==4.66.1
traitlets==5.14.0
twilio==8.11.0
typing_extensions==4.9.0
urllib3==2.0.7
wcwidth==0.2.12
websockets==12.0
yarl==1.9.4
python-multipart
uvicorn
python-dotenv